requirements.txt
# Core requirements with fixed versions
pyspark==3.1.3
confluent_kafka==2.0.2
python-dotenv==0.21.1
pandas==1.5.3
pyarrow==10.0.1

# GCP dependencies
google-cloud-storage==2.7.0
google-cloud-bigquery==3.3.5

# Data generation
faker==15.3.4

# Extras
tenacity==8.2.2